Consolidate RB, touch up pyquil utils (#426)
* Removes global noise scaling, global seeding, and unused utility functions. Use executor + unitary folding in tests.

* Move executor to pyquil_utils.

* Use existing rb_circuits function for pyquil rb_circuits. Move this from pyquil_utils to testing.

* Remove old benchmarking code and tests.

* Formatting.

* Fix docs test.

* Remove unused add_depolarizing_noise pyquil utility. Rename test file.
