+++ mitiq/benchmarks/randomized_benchmarking.py
@@ -58,21 +58,14 @@ def rb_circuits(
         qubit1 = LineQubit(qid0)
         if n_qubits == 1:
             rb_circuits = [
-                _random_single_q_clifford(
-                    qubit1,
-                    num,
-                    C1,
-                    CFD_MAT_1Q,
-                )
+                _random_single_q_clifford(qubit1, num, C1, CFD_MAT_1Q,)
                 for _ in range(trials)
             ]
         elif n_qubits == 2:
             qid1 = qubit_labels[1] if qubit_labels else 1
             qubit2 = LineQubit(qid1)
             cfd_matrices = _two_qubit_clifford_matrices(
-                qubit1,
-                qubit2,
-                CLIFFORDS,  # type: ignore
+                qubit1, qubit2, CLIFFORDS,  # type: ignore
             )
             rb_circuits = [
                 _random_two_q_clifford(
