+++ mitiq/mitiq_pyquil/tests/test_pyquil_utils.py
@@ -17,7 +17,7 @@
 from pyquil import get_qc, Program
 from pyquil.gates import X
 
-from mitiq.mitiq_pyquil.executor import (
+from mitiq.mitiq_pyquil.pyquil_utils import (
     generate_qcs_executor,
     ground_state_expectation,
 )
@@ -26,7 +26,7 @@ QVM = get_qc("2q-qvm")
 
 
 def test_ground_state_executor():
-    executor = generate_qcs_executor(QVM, ground_state_expectation, debug=True)
+    executor = generate_qcs_executor(QVM, ground_state_expectation)
     program = Program(X(0))
     assert 0.0 == executor(program)
